# v2/api/swagger.yml
openapi: 3.0.0
info:
  description: |
    This Hermez node API is the layer that allows 3rd party apps and services to interface with the node to use the layer two features of the Hermez rollup.
    Example of these apps are:
    * Wallet: send L2 transactions, check balance, ...
    * Explorer: List transactions, slots, batches, ...
    * Exchange integrations

    Note that some of the interactions with the rollup must be done using the Ethereum network directly.
    Another way to integrate with the rollup is to [deploy a node](github.com/arnaubennassar/hermez-node/v2/) and connect directly to its PostgreSQL database.
    
    ### Pagination
    
    #### Usage

    All the endpoints that return a list of undefined size use pagination. Unless the opposite is explicitly said.
    In order to use pagination, three query parameters are used:
    * `fromItem`: indicates the first item to be returned. In general, this parameter shouldn't be provided in the first call to the endpoint, and use the `itemId` of the last returned item (+/-) 1, if the order is (ascending/descending).
    * `order`: all paginated items are ordered chronologically. However, the specific fields to guarantee this order depend on each endpoint. For this purpose, `itemId` is used (itemId follows ascending chronological order except for unforged L1 user transactions). If the parameter is not provided, ascending order will be used by default. 
...
    
...

    #### Reorgs and Safety

...

    ### Signatures

...

...

...

...

...
servers:
...
paths:
...
  '/batches':
    get:
      tags:
        - Explorer
      summary: Get information about forged batches.
      description: >-
        Get information about forged batches.
      operationId: getBatches
      parameters:
        - name: minBatchNum
          in: query
          required: false
          description: Include only `batchNum > minBatchNum` batches.
          schema:
            $ref: '#/components/schemas/BatchNum'
        - name: maxBatchNum
          in: query
          required: false
          description: Include only `batchNum < maxBatchNum` batches.
          schema:
            type: number
        - name: slotNum
          in: query
          required: false
          description: Include only batches that were forged within the specified slot.
          schema:
            $ref: '#/components/schemas/SlotNum'
        - name: forgerAddr
          in: query
          required: false
          description: Include only batches forged by `forgerAddr`
          schema:
            $ref: '#/components/schemas/EthereumAddress'
        - name: fromItem
          in: query
          required: false
          description: Indicates the desired first item (using the itemId property) to be included in the response.
          schema:
            type: number
        - name: order
          in: query
          required: false
          description: Order of the returned items. Batches will be ordered by increasing `batchNum`.
          schema:
            type: string
            default: ASC
            enum:
                - ASC
                - DESC
        - name: limit
          in: query
          required: false
          description: Maximum number of items to be returned.
          schema:
            type: integer
            minimum: 1
            maximum: 2049
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batches'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'                
  '/batches/{batchNum}':
    get:
      tags:
        - Explorer
      summary: Get a specific batch.
      description: >-
        Get a specific batch.
      operationId: getBatch
      parameters:
        - name: batchNum
          in: path
          description: Batch identifier.
          required: true
          schema:
            $ref: '#/components/schemas/BatchNum'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
  '/full-batches/{batchNum}':
    get:
      tags:
        - Explorer
      summary: Get a full batch.
      description: >-
        Get a specific batch, including the associated transactions. The object returned in this method can be a bit heavy. 
        If you're devloping a front end, you may consider using a combinaton of `GET /batches/{batchnum}` and `GET /history-transactions?batchNum={batchNum}`.
      operationId: getFullBatch
      parameters:
        - name: batchNum
          in: path
          description: Batch identifier
          required: true
          schema:
            $ref: '#/components/schemas/BatchNum'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullBatch'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...
components:
  schemas:
    ItemId:
      type: integer
      description: Position of the item in the DB. This is useful for pagination, but has  nothing to do with the protocol.
...
    TransactionId:
      type: string
      description: Identifier for transactions. Used for any kind of transaction (both L1 and L2). More info on how the identifiers are built [here](https://idocs.hermez.io/#/spec/architecture/db/README?id=txid)
      example: "0x00000000000001e240004700" 
...
    EthereumAddress:
      type: string
      description: "Address of an Ethereum account."
      pattern: "^0x[a-fA-F0-9]{40}$"
      example: "0xaa942cfcd25ad4d90a62358b0dd84f33b398262a" 
...
    AccountIndex:
      type: string
      description: >-
        Identifier of an account. It references the position where the account is inside the state Merkle tree.
        The identifier is built using: `hez:` + `token symbol:` + `index`
      example: "hez:DAI:4444"
    TransactionType:
      type: string
      description: Type of transaction.
      enum:
      - Exit
      - Transfer
      - Deposit
      - CreateAccountDeposit
      - CreateAccountDepositTransfer
      - DepositTransfer
      - ForceTransfer
      - ForceExit
      - TransferToEthAddr
      - TransferToBJJ
...
    TokenId:
      type: integer
      description: Identifier of a token registered in the network.
      minimum: 0
      maximum: 4294967295
      example: 98765
    BigInt:
      type: string
      description: BigInt is an integer encoded as a string for numbers that are very large.
      example: "8708856933496328593"
      pattern: "^\\d+$"
    FeeSelector:
      type: integer
      description: Index of the fee type to select, more info [here](https://idocs.hermez.io/#/spec/zkrollup/fee-table?id=transaction-fee-table).
      minimum: 0
      maximum: 256
      example: 36
    Nonce:
      type: integer
      description: Number that can only be used once per account, increments by one at each transaction.
      minimum: 0
      maximum: 1.84467440737096e+19
      example: 121
...
    BatchNum:
      type: integer
      description: Identifier of a batch. Every new forged batch increments by one the batchNum, starting at 0.
      minimum: 0
      maximum: 4294967295
      example: 5432
...
    HistoryTransaction:
      type: object
      description: Transaction of the Hermez Network
      properties:
        L1orL2:
          type: string
          enum:
            - L1
            - L2
        id:
          $ref: '#/components/schemas/TransactionId'
        itemId:
          $ref: '#/components/schemas/ItemId'
        type:
          $ref: '#/components/schemas/TransactionType'
        position:
          $ref: '#/components/schemas/TransactionPosition'
        fromAccountIndex:
          type: string
          description: >-
            Identifier of an account. It references the position where the account is inside the state Merkle tree.
            The identifier is built using: `hez:` + `token symbol:` + `index`
          example: "hez:DAI:4444"
          nullable: true
        fromHezEthereumAddress:
          type: string
          description: "Address of an Ethereum account linked to the Hermez Network."
          pattern: "^hez:0x[a-fA-F0-9]{40}$"
          example: "hez:0xaa942cfcd25ad4d90a62358b0dd84f33b398262a"
          nullable: true
        fromBJJ:
          type: string
          description: "BabyJubJub compressed public key, encoded as base64 URL (RFC 4648), which result in 33 bytes. The padding byte is replaced by a sum of the encoded bytes."
          pattern: "^hez:[A-Za-z0-9_-]{44}$"
          example: "hez:9CK9fjQdMUTGm8KDvGLy3MB-vnP0NCcGX7Uh7OO6KRJm"
          nullable: true
        toAccountIndex:
          allOf:
            - $ref: '#/components/schemas/AccountIndex'
            - example: "hez:DAI:672"
        toHezEthereumAddress:
          type: string
          description: "Address of an Ethereum account linked to the Hermez Network."
          pattern: "^hez:0x[a-fA-F0-9]{40}$"
          example: "hez:0xaa942cfcd25ad4d90a62358b0dd84f33b398262a"
          nullable: true
        toBJJ:
          type: string
          description: "BabyJubJub compressed public key, encoded as base64 URL (RFC 4648), which result in 33 bytes. The padding byte is replaced by a sum of the encoded bytes."
          pattern: "^hez:[A-Za-z0-9_-]{44}$"
          example: "hez:f1J78_6uqTyjX6jrVCqN4RFeRBnWQAGl477ZFtOnH6Sm"
          nullable: true
        amount:
          allOf:
            - $ref: '#/components/schemas/BigInt'
            - description: Amount of tokens to be sent.
            -  example: "4903020000000000000"
        batchNum:
          type: integer
          description: Batch in which the transaction was forged. Null indicates not forged yet.
          minimum: 0
          maximum: 4294967295
          example: 5432
          nullable: true
        historicUSD:
          type: number
          description: Value in USD at the moment the transaction was forged.
          example: 49.7
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: In the case of L1 indicates the moment where the transaction was added in the smart contract. For L2 indicates when the transaction was forged.
        token:
          $ref: '#/components/schemas/Token'
        L1Info:
          type: object
          description: Additional information that only applies to L1 transactions.
          nullable: true
          properties:
            toForgeL1TransactionsNum:
              $ref: '#/components/schemas/ToForgeL1TransactionsNum'
            userOrigin:
              type: boolean
              description: True if the transaction was sent by a user. False if it was sent by a coordinator.
            depositAmount:
              allOf:
                - $ref: '#/components/schemas/BigInt'
                - description: Tokens transfered from L1 to L2.
                - example: "4900000000000000000"
            amountSuccess:
              type: boolean
              description: >-
                Indicates if the amount specified by the user has been sent propperly. If false, the amount that actualy has been sent is 0.
                If the transaction hasn't been forged yet (`batchNum == null`), this value will be false, as it's unknown if it has succeed or not.
                An example were this value could be false: a user send a `DepositTransfer` transaction, but when the tx is forged there are not
                enougth founds in the account. In this case the transfer part of the transaction won't be effective making the amount have a real value of 0.
              example: true
            depositAmountSuccess:
              type: boolean
              description: >-
                Indicates if the deposit amount specified by the user has been sent propperly. If false, the deposit amount that actualy has been sent is 0.
                If the transaction hasn't been forged yet (`batchNum == null`), this value will be false, as it's unknown if it has succeed or not.
                An example were this value could be false: a user send a `Deposit` transaction, but when the tx is forged the token id is not registered on the network.
                In this case transaction won't be effective making the deposit amount have a real value of 0.
              example: true
            historicDepositAmountUSD:
              type: number
              description: Deposit amount in USD, at the moment the transaction was made.
              example: 3.897
              nullable: true
            ethereumBlockNum:
              allOf:
                - $ref: '#/components/schemas/EthBlockNum'
                - description: Ethereum block in which the transaction was added to the smart contract forge queue.
                - example: 258723049
          required:
            - toForgeL1TransactionsNum
            - userOrigin
            - depositAmount
            - amountSuccess
            - depositAmountSuccess
            - historicDepositAmountUSD
            - ethereumBlockNum
          additionalProperties: false
        L2Info:
          type: object
          description: Additional information that only applies to L2 transactions.
          nullable: true
          properties:
            fee:
              $ref: '#/components/schemas/FeeSelector'
            historicFeeUSD:
              type: number
              description: Fee in USD, at the moment the transaction was forged.
              example: 263.89
              nullable: true
            nonce:
              $ref: '#/components/schemas/Nonce'
          example: null
          required:
            - fee
            - historicFeeUSD
            - nonce
          additionalProperties: false
      example:
        L1Info:
          ethereumBlockNum: 1
          historicDepositAmountUSD: 232.47
          depositAmount: '261'
          toForgeL1TransactionsNum: 10
          userOrigin: true
        L1orL2: L1
        L2Info: 
        amount: '261'
        batchNum: 1
        fromAccountIndex: hez:ETH:276
        fromBJJ: hez:p_OohTzjzZnD3Sw93HQlK13DSxfD6lyvbfhh2kBsV6Z4
        fromHezEthereumAddress: hez:0x0000000000000000000000000000000000000114
        historicUSD: 3784.19
        id: '0x00000000000000000a000400'
        itemId: 2
        position: 4
        timestamp: '2020-11-17T14:08:12.197554Z'
        toAccountIndex: hez:ETH:276
        toBJJ: hez:p_OohTzjzZnD3Sw93HQlK13DSxfD6lyvbfhh2kBsV6Z4
        toHezEthereumAddress: hez:0x0000000000000000000000000000000000000114
        token:
          USD: 234.56
          decimals: 18
          ethereumAddress: '0x0000000000000000000000000000000000000000'
          ethereumBlockNum: 0
          fiatUpdate: 
          id: 0
          itemId: 1
          name: Ether
          symbol: ETH
        type: CreateAccountDeposit
      required:
        - L1orL2
        - id
        - itemId
        - type
        - position
        - fromAccountIndex
        - fromHezEthereumAddress
        - fromBJJ
        - toAccountIndex
        - toHezEthereumAddress
        - toBJJ
        - amount
        - batchNum
        - historicUSD
        - timestamp
        - token
        - L1Info
        - L2Info
      additionalProperties: false
...
    EthBlockNum:
      type: integer
      description: Ethereum block number
      minimum: 0
      maximum: 1.84467440737096e+19
      example: 762375478
    ToForgeL1TransactionsNum:
      type: integer
      description: Reference to know in which batch a L1 transaction was forged / will be forged.
      minimum: 0
      maximum: 4294967295
      example: 784
      nullable: true
    TransactionPosition:
      type: integer
      description: Position that a transaction occupies in a batch.
      minimum: 0
      example: 5
...
    TokenSymbol:
      type: string
      description: Abreviation of the token name.
      example: "DAI"
...
    CollectedFees:
      type: object
      description: Collected fees by the forger of the batch, represented by a map of tokenId => amount. A maximum of 64 different tokens can be used.
      additionalProperties:
        type: string
      example:
        1234: "425632785672345647"
        4321: "86538967235465432654352"
    Batch:
      type: object
      description: Group of transactions forged in a coordinator and sent and validated in Ethereum.
      properties:
        itemId:
          $ref: '#/components/schemas/ItemId'
        batchNum:
          $ref: '#/components/schemas/BatchNum'
        ethereumBlockNum:
          $ref: '#/components/schemas/EthBlockNum'
        ethereumBlockHash:
          type: string
          description: hash of the Ethereum block in which the batch was forged
          example: "0xfe88c94d860f01a17f961bf4bdfb6e0c6cd10d3fda5cc861e805ca1240c58553"
        ethereumTxHash:
          type: string
          description: hash of the Ethereum transaction in which the batch was forged
          example: "0xae98c94d860f01a17f961bf4bdfb6e0c6cd10d3fda5cc861e805ca1240c58553"
        timestamp:
          type: string
          format: date-time
          description: Time in which the batch was forged.
        forgerAddr:
          $ref: '#/components/schemas/EthereumAddress'
        collectedFees:
          $ref: '#/components/schemas/CollectedFees'
        historicTotalCollectedFeesUSD:
          type: number
          description: Sum of the all the fees collected, in USD, at the moment the batch was forged.
          example: 23.3
        stateRoot:
          allOf:
            - $ref: '#/components/schemas/Hash'
            - description: Root of the accounts Merkle Tree.
            - example: "2734657026572a8708d883"
        numAccounts:
          type: integer
          description: Number of registered accounts in this batch.
        exitRoot:
          allOf:
            - $ref: '#/components/schemas/Hash'
            - description: Root of the exit Merkle Tree associated to this batch.
            - example: "2734657026572a8708d883"
        forgeL1TransactionsNum:
          type: integer
          description: Identifier that corresponds to the group of L1 transactions forged in the current batch.
          example: 5
          nullable: true
        slotNum:
          $ref: '#/components/schemas/SlotNum'
        forgedTransactions:
          type: integer
          description: Amount of forged transactions in this batch.
          example: 318
      additionalProperties: false
      required:
          - itemId
          - batchNum
          - ethereumTxHash
          - ethereumBlockNum
          - ethereumBlockHash
          - timestamp
          - forgerAddr
          - collectedFees
          - historicTotalCollectedFeesUSD
          - stateRoot
          - numAccounts
          - exitRoot
          - forgeL1TransactionsNum
          - slotNum
          - forgedTransactions
    FullBatch:
      type: object
      description: Group of transactions forged in a coordinator and sent and validated in Ethereum.
      properties:
        batch:
          $ref: '#/components/schemas/Batch'
        transactions:
          type: array
          description: List of forged transactions in the batch
          items:
            $ref: '#/components/schemas/HistoryTransaction'
          nullable: true
      additionalProperties: false
      required:
        - batch
        - transactions
    Hash:
      type: string
      description: hashed data
      example: "2734657026572a8708d883"
    SlotNum:
      type: integer
      description: Identifier of a slot.
      minimum: 0
      maximum: 4294967295
      example: 784        
    Batches:
      type: object
      properties:
        batches:
          type: array
          description: List of batches.
          items:
            $ref: '#/components/schemas/Batch'
        pendingItems:
          $ref: '#/components/schemas/PendingItems'
      additionalProperties: false
      required:
        - batches
        - pendingItems
...
    Token:
      type: object
      description: Hermez Network compatible and registered token.
      properties:
        id:
          $ref: '#/components/schemas/TokenId'
        ethereumAddress:
          allOf:
            - $ref: '#/components/schemas/EthereumAddress'
            - description: Ethereum address in which the token is deployed.
            - example: "0xaa942cfcd25ad4d90a62358b0dd84f33b398262a"
        itemId:
          $ref: '#/components/schemas/ItemId'
        name:
          type: string
          description: full name of the token
          example: Maker Dai
        symbol:
          allOf:
            - $ref: '#/components/schemas/TokenSymbol'
            - example: DAI
        decimals:
          type: integer
          description: Number of decimals of the token.
          example: 18
        ethereumBlockNum:
          allOf:
            - $ref: '#/components/schemas/EthBlockNum'
            - description: Ethereum block number in which the token was added to the Hermez Network.
            - example: 539847538
        USD:
          type: number
          description: Value of the token in USD.
          example: 1.01
          nullable: true
        fiatUpdate:
          type: string
          format: date-time
          description: Timestamp of the moment the `USD` value was updated.
          nullable: true
      required:
        - id
        - ethereumAddress
        - itemId
        - name
        - symbol
        - decimals
        - ethereumBlockNum
        - USD
        - fiatUpdate
      additionalProperties: false
...
    PendingItems:
      type: integer
      description: Amount of items that will be returned in subsequent calls to the endpoint, as long as they are done with same filters. When the value is 0 it means that all items have been sent.
      example: 15
...
    Error:
      type: object
      description: Error response. 
      properties:
        message:
          type: string
        code:
          type: integer
          enum: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22]
          description: >
            Error codes:
              * `1` - ErrParamValidationFailed error
              * `2` - ErrDuplicatedKey error
              * `3` - ErrSQLTimeout error
              * `4` - ErrSQLNoRows error
              * `5` - ErrExitAmount0 error
              * `6` - ErrInvalidTxTypeOrTxID error
              * `7` - ErrFeeOverflow error
              * `8` - ErrGettingSenderAccount error
              * `9` - ErrGettingSenderAccount error
              * `10` - ErrInvalidNonce error
              * `11` - ErrInvalidSignature error
              * `12` - ErrGettingReceiverAccount error
              * `13` - ErrCantSendToEthAddr error
              * `14` - ErrNotAtomicTxsInPostPoolTx error
              * `15` - ErrFailedToGetCurrentBlock error
              * `16` - ErrFailedToGetAuctionVars error
              * `17` - ErrFailedToAddEmptySlot error
              * `18` - ErrTxsNotAtomic error
              * `19` - ErrSingleTxInAtomicEndpoint error
              * `20` - ErrRqOffsetOutOfBounds error
              * `21` - ErrInvalidAtomicGroupID error
              * `22` - ErrFailedToFindOffsetToRelativePositionCode error
        type:
          type: string
          enum:
            - ErrParamValidationFailed
            - ErrDuplicatedKey
            - ErrSQLTimeout
            - ErrSQLNoRows
            - ErrExitAmount0
            - ErrInvalidTxTypeOrTxID
            - ErrFeeOverflow
            - ErrGettingSenderAccount
            - ErrAccountTokenNotEqualTxToken
            - ErrInvalidNonce
            - ErrInvalidSignature
            - ErrGettingReceiverAccount
            - ErrCantSendToEthAddr
            - ErrNotAtomicTxsInPostPoolTx
            - ErrFailedToGetCurrentBlock
            - ErrFailedToGetAuctionVars
            - ErrFailedToAddEmptySlot
            - ErrTxsNotAtomic
            - ErrSingleTxInAtomicEndpoint
            - ErrInvalidAtomicGroupID
            - ErrFailedToFindOffsetToRelativePosition
          description: >
            Error types:
              * `ErrParamValidationFailed` - this error appears when user sends wrong info in query|uri|form params
              * `ErrDuplicatedKey` - record already exists in db
              * `ErrSQLTimeout` - node is under heavy pressure and sql connection timeout happened
              * `ErrSQLNoRows` - there is no such item in database
              * `ErrExitAmount0` - receiving (and rejecting) a tx of type exit with amount 0 in /transactions-pool request
              * `ErrInvalidTxTypeOrTxID` - tx with invalid type or txID was sent in /transactions-pool request
              * `ErrFeeOverflow` - tx with fee amount that doesn't fit in 128 bits was sent in /transactions-pool request
              * `ErrGettingSenderAccount` - sender account for tx was not found in /transactions-pool request
              * `ErrAccountTokenNotEqualTxToken` - account token is not equal to tx token in /transactions-pool request
              * `ErrInvalidNonce` - invalid nonce was sent in /transactions-pool request
              * `ErrInvalidSignature` - invalid signature was sent in /transactions-pool or /account-creation-authorization requests
              * `ErrGettingReceiverAccount` - receiver account for tx was not fount in /transactions-pool request
              * `ErrCantSendToEthAddr` - it's not possible to send to Eth address when there is no account index associated to this Eth address and token, or the coordinator hasn't got an authorization to create an account
              * `ErrNotAtomicTxsInPostPoolTx` - user sending atomic tx to /transactions-pool request
              * `ErrFailedToGetCurrentBlock` - can't get current block in /slots request
              * `ErrFailedToGetAuctionVars` - can't get auction vars in /slots request
              * `ErrFailedToAddEmptySlot` - can't add empty slot in /slots request
              * `ErrTxsNotAtomic` - there is at least one transaction in the payload that could be forged without the others
              * `ErrSingleTxInAtomicEndpoint` - to use the atomic-pool endpoint at least two transactions are required
              * `ErrRqOffsetOutOfBounds` - one of the transactions requested another one outside the bounds of the provided array
              * `ErrInvalidAtomicGroupID` - invalid atomicGroupId
              * `ErrFailedToFindOffsetToRelativePosition` - can't find offset to relative position
    Error400:
      allOf:
        - $ref: '#/components/schemas/Error'
        - example:
            message: Invalid signature.
            code: 4
            type: ErrInvalidSignature
    Error404:
      allOf:
        - $ref: '#/components/schemas/Error'
        - example:
            message: Item(s) not found.
    Error500:
      allOf:
        - $ref: '#/components/schemas/Error'
        - example:
            message: Database error.
...